# app/backend/docs/swagger.yaml
definitions:
...
  dto.CreateProfessorDto:
    properties:
      hoursToAllocate:
        maximum: 3
        minimum: 1
        type: integer
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - hoursToAllocate
    - name
    type: object
...
  dto.UpdateProfessorDto:
    properties:
      hoursToAllocate:
        maximum: 3
        minimum: 1
        type: integer
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - hoursToAllocate
    - name
    type: object
...
  httperr.Fields:
    properties:
      field:
        type: string
      message:
        type: string
      value: {}
    type: object
  httperr.RestErr:
    properties:
      code:
        type: integer
      error:
        type: string
      fields:
        items:
          $ref: '#/definitions/httperr.Fields'
        type: array
      message:
        type: string
    type: object
...
  response.ManyProfessorsResponse:
    properties:
      professors:
        items:
          $ref: '#/definitions/response.ProfessorResponse'
        type: array
    type: object
...
  response.ProfessorResponse:
    properties:
      hoursToAllocate:
        type: integer
      name:
        type: string
      uuid:
        type: string
    type: object
...
info:
  contact:
    email: your@mail.com
    name: API Support
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Time Table API
  version: "1.0"
paths:
...
  /professors:
    post:
      consumes:
      - application/json
      description: Endpoint for create professor
      parameters:
      - description: Create professor dto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProfessorDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Create new professor
      tags:
      - professor
  /professors/{uuid}:
    delete:
      consumes:
      - application/json
      description: delete professor by uuid
      parameters:
      - description: professor uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Delete professor
      tags:
      - professor
    get:
      consumes:
      - application/json
      description: Get professor by uuid
      parameters:
      - description: professor uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProfessorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Professor details
      tags:
      - professor
    patch:
      consumes:
      - application/json
      description: Endpoint for update professor
      parameters:
      - description: professor uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: Update professor dto
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.UpdateProfessorDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Update professor
      tags:
      - professor
  /professors/list-all:
    get:
      consumes:
      - application/json
      description: Get many professors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ManyProfessorsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Get many professors
      tags:
      - professor
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"